libopencad: rename io.h to avoid conflict with header from mingw compiler
Since the update to mingw-w64-x86_64-headers-14.0.0.r302.gd7f3c5201-1
the build fails with:

```
In file included from D:/a/_temp/msys64/mingw64/include/sys/stat.h:14,
                 from D:/a/_temp/msys64/mingw64/include/wchar.h:19,
                 from D:/a/_temp/msys64/mingw64/include/c++/16.2.0/cwchar:49,
                 from D:/a/_temp/msys64/mingw64/include/c++/16.2.0/bits/postypes.h:42,
                 from D:/a/_temp/msys64/mingw64/include/c++/16.2.0/bits/char_traits.h:44,
                 from D:/a/_temp/msys64/mingw64/include/c++/16.2.0/string:45,
                 from ..\..\..\..\..\ogr\ogrsf_frmts\cad\libopencad/cadheader.h:20,
                 from ..\..\..\..\..\ogr\ogrsf_frmts\cad\libopencad/cadobjects.h:19,
                 from ..\..\..\..\..\ogr\ogrsf_frmts\cad\libopencad/cadgeometry.h:18,
                 from ..\..\..\..\..\ogr\ogrsf_frmts\cad\libopencad\dwg\r2000.cpp:15:
..\..\..\..\..\ogr\ogrsf_frmts\cad\libopencad\dwg/io.h:147:5: error: 'CADHandle' does not name a type
  147 |     CADHandle ReadHANDLE();
```

This is due to a confusion between io.h from the compiler header with
the one from libopencad
