+++ ogr/ogrsf_frmts/cad/libopencad/dwg/dwg_io.h
